file_services/requirements.txt
boto3
fastapi
uvicorn
mistralai
python-dotenv
httpx
pydantic
python-multipart 